+++ src/core/fetch.ts
@@ -21,6 +21,10 @@ export const COOKIE_CONSENT_GRAPHQL_QUERY = /* GraphQL */ `
       expiryDays
       cookieDomain
       position
+      colorBackground
+      colorText
+      colorAccent
+      colorAccentText
     }
   }
 `;